pyproject.toml
[project]
name = "brief-ai-vs-software-engineering"
version = "0.1.0"
description = "Build a concrete and critical brief comparing Software Engineering vs AI Engineering."
requires-python = ">=3.10"
dependencies = [
  "openai>=1.60.0",
  "python-dotenv>=1.0.1",
  "pytest>=8.3.0",
]

[tool.uv]
package = false

[build-system]
requires = ["setuptools>=68.0.0", "wheel"]
build-backend = "setuptools.build_meta"
